# MusicAPI/openapi.yml
openapi: '3.0.3'
info:
  title: Music Social Media API
  description: Rest API for a social media platform with ML-based Content/Collaborative/Hybrid User Specific Recommendations
  version: '1.0'
servers:
  - url: https://mt241063-10974.node.ustp.cloud
    description: Deployment
  - url: http://localhost:3000/
    description: Local Testing
paths:
...
  /interactions/posts/{post_id}/like:
    post:
      summary: "Like a specific post"
      tags: [ Interactions ]
      security:
        - BearerAuth: []
      parameters:
        - name: post_id
          in: path
          required: true
          schema:
            type: string

      responses:
        "200":
          description: "Liked successfully"

        "400":
          description: "Bad Request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
              example:
                message: "You have already liked this post"
                code: "ALREADY_LIKED"

        "401":
          description: "Unauthorized"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
              example:
                message: "You need to be logged in in order to like a post."
                code: "NOT_LOGGED_IN"
  /interactions/posts/{post_id}/view:
    post:
      summary: "Sent to API in order to document watchtime"
      tags: [ Interactions ]
      security:
        - BearerAuth: [ ]
      parameters:
        - name: post_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                watch_time_seconds:
                  type: integer
                  description: "Seconds the user has watched the post"
                  example: 36
                completed:
                  type: boolean
                  description: "The post was viewed for >= 90% of the post-duration"
      responses:
        "200":
          description: "Watchtime stored successfully"

        "400":
          description: "Bad Request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
              example:
                message: "Missing fields"
                code: "MISSING_FIELDS"

        "401":
          description: "Unauthorized"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
              example:
                message: "You need to be logged in in order to store watchtime"
                code: "NOT_LOGGED_IN"
  /interactions/posts/{post_id}/comment:
    post:
      summary: "Comment on a post by logged in user"
      tags: [ Interactions ]
      security:
        - BearAuth: []
      parameters:
        - name: post_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                comment:
                  type: string
                  description: "Comment by the user"
                  example: "Great Song!"

      responses:
        "200":
          description: "Watchtime stored successfully"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentResponse"

        "400":
          description: "Bad Request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
              example:
                message: "Missing fields. Comment field is probably empty"
                code: "MISSING_FIELDS"
        "401":
          description: "Unauthorized"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
              example:
                message: "You need to be logged in in order to add a comment"
                code: "NOT_LOGGED_IN"
  /post:
    post:
      summary: "Upload a post"
      tags: [ Posts ]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/PostUpload"
      responses:
        "200":
          description: "Post uploaded successfully"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostResponse"
        "400":
          description: "Bad Request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
              examples:
                ImageTooLarge:
                  summary: Picture too large
                  value:
                    message: "The picture exceeds the maximum file size of 5MB"
                    code: "IMAGE_TOO_LARGE"
                AudioTooLarge:
                  summary: Audio too large
                  value:
                    message: "The audio file exceeds the maximum file size of 5MB"
                    code: "AUDIO_TOO_LARGE"
                MissingFields:
                  summary: Missing fields
                  value:
                    message: "Missing required fields: post_title, post_audio"
                    code: "MISSING_FIELDS"
        "401":
          description: "Unauthorized"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
              example:
                message: "Session expired please log in"
                code: "TOKEN_EXPIRED"
  /post/{post_id}:
    get:
      summary: "Get the Post for the user to view"
      tags: [ Posts ]
      parameters:
        - name: post_id
          in: path
          required: true
          schema: 
            type: string
      responses:
        "200":
          description: "Post fetched successfully"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostResponse"
        "400":
          description: "Bad Request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
              example:
                message: "Invalid post ID (Not a Number or 0)"
                code: "INVALID_POST_ID"
        "404":
          description: "Not Found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
              example:
                message: "Post not found. Check Post ID"
                code: "POST_NOT_FOUND"

    patch:
      summary: "Change values of a already existing post"
      tags: [ Posts ]
      security:
        - BearerAuth: [ ]
      parameters:
        - name: post_id
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/PostUpdate"
      responses:
        "200":
          description: "Post updated successfully"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostResponse"
        "400":
          description: "Bad Request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
              examples:
                ImageTooLarge:
                  summary: Picture too large
                  value:
                    message: "The picture exceeds the maximum file size of 5MB"
                    code: "IMAGE_TOO_LARGE"
                InvalidInput:
                  summary: Invalid input
                  value:
                    message: "This is not a valid input."
                    code: "INVALID_INPUT"
        "401":
          description: "Unauthorized"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
              example:
                message: "Session expired please log in"
                code: "TOKEN_EXPIRED"
  /post/{post_id}/comments:
    get:
      summary: "Loads all the comments for a specific post"
      tags: [ Posts ]
      parameters:
        - name: post_id
          in: path
          schema: 
            type: string
      responses:
        "200":
          description: "Comments fetched successfully"
          content:
            application/json:
              schema:
                type: array
                items: 
                  $ref: "#/components/schemas/CommentResponse"
        "400":
          description: "Bad Request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
              examples:
                InvalidInput:
                  summary: Invalid input
                  value:
                    message: "Input is not a number or 0"
                    code: "INVALID_INPUT"
components:
  schemas:
...
    PostBase:
      type: object
      required: [ post_title, post_description ]
      properties:
        post_title:
          type: string
          description: "Title for the Post"
          example: "Great Post Title"

        post_description:
          type: string
          description: "Description and additional infos to the post"
    PostUpdateableParameters:
      type: object
      properties:
        post_image:
          type: string
          format: binary
          description: "Image for the Post"

        post_tags:
          type: array
          items:
            type: string
          description: "Tags to be associated with the post"
          example: [ "chill", "dnb", "great culture" ]

        post_audio_genres:
          type: array
          description: "Genre of the post audio. Separate from post_tags"
          items:
            type: string
          example: [ "Reggae","Progressive Rock" ]
    PostUpload:
        allOf:
          - $ref: "#/components/schemas/PostBase"
          - $ref: "#/components/schemas/PostUpdateableParameters"
          - type: object
            required: [post_audio, post_image, post_audio_genres]
            properties:

              post_audio:
                type: string
                format: binary
                description: "Audio for the Post"
    PostUpdate:
      minProperties: 1
      allOf:
        - type: object
          properties:
           post_title:
              type: string
              description: "Title for the Post"
              example: "Great Post Title"

           post_description:
              type: string
              description: "Description and additional infos to the post"
        - $ref: "#/components/schemas/PostUpdateableParameters"
    PostResponse:
      allOf:
        - $ref: "#/components/schemas/PostBase"
        - type: object
          properties:
            post_id:
              type: integer
              description: "ID of a specific post"
              example: 276
            post_audio_url:
              type: string
              format: url
              description: "Relative path to the Audio file for the post"
              example: "/uploads/post_id/audio.mp3"
            post_image_url:
              type: string
              format: url
              description: "Relative path to the Image file for the post"
              example: "/uploads/post_id/image.jpg"
            post_views_count:
              type: integer
              description: "Number of views on that post"
              example: "123"
            post_likes_count:
              type: integer
              description: "Number of likes on that post"
              example: "24"
            post_comments_count:
              type: integer
              description: "Number of comments on that post"
              example: "3"
            post_audio_genres:
              type: array
              description: "Audio Genres that were associated when the post was created"
              example: ["Reggae", "Rock"]
            post_tags:
              type: array
              description: "Tags that are associated when the post was created. May be empty!"
              example: ["Tag", "Tag2"]
...
    CommentResponse:
      type: object
      properties:
        username:
          type: string
          description: "Username of a Profile owner"
          example: "Mike"
        comment:
          type: string
          description: "Comment made by the user"
          example: "Great post!"
        comment_time:
          type: string
          format: date-time
          description: "Time when the comment was posted"
          example: "2026-01-23T18:15:00Z"
...
    ErrorMessage:
      type: object
      properties:
        message:
          type: string
          example: "Wrong credentials"
        code:
          type: string
          example: "INVALID_CREDENTIALS"
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT